+++ iot-storage-client/iot/storage/client/_client.py
@@ -15,6 +15,11 @@ from azure.storage.blob import (
     generate_blob_sas,
 )
 
+from ._helpers import (
+    generate_cloud_conn_str,
+    generate_edge_conn_str,
+    generate_local_conn_str,
+)
 from ._types import CredentialType, LocationType
 
 
@@ -25,7 +30,9 @@ class IoTStorageClient:
     location_type: str
     account_name: str
     credential: str
+    connection_string: str
 
+    module: Optional[str] = None
     host: Optional[str] = None
     port: Optional[str] = None
 
@@ -39,6 +46,7 @@ class IoTStorageClient:
         location_type: str,
         account_name: str,
         credential: str,
+        module: Optional[str] = None,
         host: Optional[str] = None,
         port: Optional[str] = None,
     ) -> None:
@@ -46,6 +54,7 @@ class IoTStorageClient:
         self.location_type = location_type
         self.account_name = account_name
         self.credential = credential
+        self.module = module
         self.host = host
         self.port = port
         self.instantiate_service_client()
@@ -57,43 +66,72 @@ class IoTStorageClient:
             f"credential type: {self.credential_type.lower()}\n"
             f"location type: {self.location_type.lower()}\n"
             f"account name: {self.account_name}\n"
-            f"credential: {self.credential[0:5]}****"
+            f"credential: {self.credential[0:3]}****"
         )
 
     def instantiate_service_client(self) -> None:
         """init service_client based on credential and location types"""
-        if self.credential_type == CredentialType.CONNECTION_STRING:
+        if self.credential == CredentialType.CONNECTION_STRING:
+            self.connection_string = self.credential
             self.service_client = BlobServiceClient.from_connection_string(
                 self.credential
             )
-        elif self.credential_type == CredentialType.ACCOUNT_KEY:
-            if self.location_type == LocationType.CLOUD_BASED:
-                self.service_client = BlobServiceClient(
-                    account_url=f"https://{self.account_name}.blob.core.windows.net",
-                    credential=self.credential,
-                )
-            elif self.location_type == LocationType.EDGE_BASED:
-                self.service_client = BlobServiceClient(
-                    account_url=f"http://{self.host}:{self.port}/{self.account_name}.blob.core.windows.net",
-                    credential=self.credential,
-                )
-            else:
-                self.location_type = "INVALID"
-                print(
-                    "invalid location type, please use one of LocationType[CLOUD_BASED, EDGE_BASED]"
-                )
         else:
-            self.credential_type = "INVALID"
-            print(
-                "invalid credential type, please use one of CredentialType[ACCOUNT_KEY, CONNECTION_STRING"
-            )
+            if self.location_type == LocationType.CLOUD_BASED:
+                connection_string = generate_cloud_conn_str(
+                    account=self.account_name,
+                    account_key=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_KEY
+                    else None,
+                    account_sas=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_SAS
+                    else None,
+                )
+                self.connection_string = connection_string
+                self.service_client = BlobServiceClient.from_connection_string(
+                    connection_string
+                )
+            if self.location_type == LocationType.EDGE_BASED:
+                connection_string = generate_edge_conn_str(
+                    host=self.host,
+                    port=self.port,
+                    account=self.account_name,
+                    account_key=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_KEY
+                    else None,
+                    account_sas=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_SAS
+                    else None,
+                )
+                self.connection_string = connection_string
+                self.service_client = BlobServiceClient.from_connection_string(
+                    connection_string
+                )
+            if self.location_type == LocationType.LOCAL_BASED:
+                connection_string = generate_local_conn_str(
+                    module=self.module,
+                    port=self.port,
+                    account=self.account_name,
+                    account_key=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_KEY
+                    else None,
+                    account_sas=self.credential
+                    if self.credential_type == CredentialType.ACCOUNT_SAS
+                    else None,
+                )
+                self.connection_string = connection_string
+                self.service_client = BlobServiceClient.from_connection_string(
+                    connection_string
+                )
 
     def format_account_url(self) -> str:
         """format the blob account url"""
         if self.location_type == LocationType.CLOUD_BASED:
             return f"https://{self.account_name}.blob.core.windows.net"
         if self.location_type == LocationType.EDGE_BASED:
-            return f"http://{self.host}:{self.port}/{self.account_name}.blob.core.windows.net"
+            return f"http://{self.host}:{self.port}/{self.account_name}"
+        if self.location_type == LocationType.LOCAL_BASED:
+            return f"http://{self.module}:{self.port}/{self.account_name}"
         return "invalid LocationType"
 
     def container_exists(self, container_name: str) -> bool:
@@ -420,10 +458,7 @@ class IoTStorageClient:
     ) -> Union[str, None]:
         """generate a SAS URL for a given file inside the container"""
         try:
-            account_key = self.credential
-            if self.credential_type != CredentialType.ACCOUNT_KEY:
-                account_key = self.service_client.credential.account_key
-
+            account_key = self.service_client.credential.account_key
             sas_token = generate_blob_sas(
                 account_name=self.account_name,
                 container_name=container_name,