+++ iot-storage-client/iot/storage/client/_types.py
@@ -5,6 +5,7 @@ class CredentialType:
     """iot storage client credential types for authentication"""
 
     ACCOUNT_KEY = "ACCOUNT_KEY"
+    ACCOUNT_SAS = "ACCOUNT_SAS"
     CONNECTION_STRING = "CONNECTION_STRING"
 
 
@@ -13,3 +14,4 @@ class LocationType:
 
     CLOUD_BASED = "CLOUD_BASED"
     EDGE_BASED = "EDGE_BASED"
+    LOCAL_BASED = "LOCAL_BASED"